+++ src/database/queries/funders/selectWithPagination.sql
@@ -17,17 +17,23 @@ WITH
 		FROM funders
 			CROSS JOIN search_query
 		WHERE
-			search_query.tsquery IS NULL
-			OR funders.name_search @@ search_query.tsquery
-			OR EXISTS (
-				SELECT 1
-				FROM funder_collaborative_members AS fcm
-					INNER JOIN funders AS member_f
-						ON fcm.member_funder_short_code = member_f.short_code
-				WHERE
-					fcm.funder_collaborative_short_code = funders.short_code
-					AND NOT is_expired(fcm.not_after)
-					AND member_f.name_search @@ search_query.tsquery
+			(
+				:isCollaborative::boolean IS NULL
+				OR funders.is_collaborative = :isCollaborative::boolean
+			)
+			AND (
+				search_query.tsquery IS NULL
+				OR funders.name_search @@ search_query.tsquery
+				OR EXISTS (
+					SELECT 1
+					FROM funder_collaborative_members AS fcm
+						INNER JOIN funders AS member_f
+							ON fcm.member_funder_short_code = member_f.short_code
+					WHERE
+						fcm.funder_collaborative_short_code = funders.short_code
+						AND NOT is_expired(fcm.not_after)
+						AND member_f.name_search @@ search_query.tsquery
+				)
 			)
 	),