+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `GET /funders` now supports the `isCollaborative` query parameter to filter funders by their collaborative status.
+
 ## 0.33.0 2026-04-14
 
 ### Added